pyproject.toml
[project]
name = "deeplearning-project-test-real-time"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "mediapipe==0.10.5",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "timm>=1.0.25",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"
